requirements.txt
tensorflow==2.15.0
numpy==1.24.3
matplotlib==3.8.0
Pillow==10.1.0
streamlit>=1.28.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
pandas==2.1.3
seaborn==0.13.0
watchdog